test: refactore fixtures to better match implementation
